+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Emails are now obfuscated in logs.
 - Onboarding shows more obviously it's possible to use manual accounts with Kresus.
 - Recurring transactions are now available from the lateral menu.
+- One can now set the balance of accounts in disabled accesses.
 
 ### Bug fixes